requirements.txt
matplotlib
numpy
pandas
seaborn
empath
gensim
nltk
scipy
scikit-learn
pyspellchecker
transformers
vaderSentiment
tqdm